pyproject.toml
[project]
name = "mesh-to-sim-asset"
version = "0.1.6"
description = "Pipeline for converting mesh files into simulation assets"
readme = "README.md"
requires-python = ">=3.10, <3.11"
license = {text = "MIT"}
authors = [{name = "Nicholas Pfaff"}]
keywords = [
    "robotics",
    "simulation",
    "drake",
    "mesh",
    "physics",
    "sdf",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Topic :: Scientific/Engineering",
]
dependencies = [
    "bpy==4.0.0",
    "coacd>=1.0.7",
    "datasets>=4.0.0",
    "drake>=1.43.0",
    "lxml>=5.4.0",
    "manipulation>=2025.5.18",
    "numpy>=2.2.6",
    "objaverse>=0.1.7",
    "openai>=1.12.0",
    "pandas>=2.3.1",
    "Pillow>=10.2.0",
    "pycollada>=0.9.2",
    "pyvista>=0.45.2",
    "scipy>=1.15.3",
    "tetgen>=0.6.5",
    "tqdm>=4.67.1",
    "trimesh>=4.6.10",
    "vhacdx>=0.0.8.post2",
]

[project.urls]
Homepage = "https://github.com/nepfaff/mesh-to-sim-asset"
Repository = "https://github.com/nepfaff/mesh-to-sim-asset"
Issues = "https://github.com/nepfaff/mesh-to-sim-asset/issues"

[dependency-groups]
dev = [
    "black==25.1.0",
    "ipython>=8.37.0",
    "isort==5.12.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["mesh_to_sim_asset"]

[tool.hatch.build.targets.wheel.force-include]
"make_asset_drake_compatible.py" = "make_asset_drake_compatible.py"
"create_drake_asset_from_geometry.py" = "create_drake_asset_from_geometry.py"

[[tool.uv.index]]
name = "blender"
url = "https://download.blender.org/pypi/"

[tool.isort]
profile = 'black'
lines_between_types = 1
combine_as_imports = true
known_first_party = ['mesh_to_sim_asset']
